perf: split MetricCard into Astro shell + tween-value island (#189)
Card shells (title, type badge, description) render from SSR HTML with
zero JS. Interactive parts are split into small islands:

- metrics-poller.tsx: invisible island, owns the BroadcastChannel tab
  leader + 10s poll, dispatches metrics:updated on the window
- metric-tween-value.tsx: one per gauge/counter card, listens for
  metrics:updated and tweens the displayed value
- Sparkline: already a standalone island, unchanged

metrics.astro now renders the full page in Astro; MetricsOverview is
deleted. On first paint all card content is visible from SSR HTML before
any island hydrates.
